Move tools to crates.
