+++ crates/core/tool/src/lib.rs
@@ -1,3 +1,9 @@
+mod errors;
+mod tool;
+
+pub use errors::*;
+pub use tool::*;
+
 use std::env;
 use std::path::Path;
