+++ crates/core/toolchain/Cargo.toml
@@ -9,11 +9,10 @@ moon_constants = { path = "../constants" }
 moon_error = { path = "../error" }
 moon_lang = { path = "../lang" }
 moon_logger = { path = "../logger" }
-moon_node_lang = { path = "../../node/lang" }
 moon_platform_runtime = { path = "../platform-runtime" }
 moon_terminal = { path = "../terminal" }
+moon_tool = { path = "../tool" }
 moon_utils = { path = "../utils" }
 proto_core = { path = "../../proto/core" }
-proto_node = { path = "../../proto/node" }
 rustc-hash = { workspace = true }
 thiserror = { workspace = true }
