+++ crates/core/toolchain/src/lib.rs
@@ -1,11 +1,7 @@
 mod errors;
-pub mod helpers;
 mod manager;
 mod toolchain;
-pub mod tools;
-mod traits;
 
 pub use errors::ToolchainError;
-pub use helpers::get_path_env_var;
+pub use moon_tool::*;
 pub use toolchain::Toolchain;
-pub use traits::{DependencyManager, RuntimeTool};
