+++ crates/core/toolchain/src/manager.rs
@@ -1,15 +1,18 @@
-use crate::{RuntimeTool, ToolchainError};
+use crate::errors::ToolchainError;
 use moon_platform_runtime::{Runtime, Version};
+use moon_tool::Tool;
 use rustc_hash::FxHashMap;
 use std::fmt::Debug;
 
+pub type CachedTool = Box<dyn Tool>;
+
 #[derive(Debug)]
-pub struct ToolManager<T: RuntimeTool + Debug> {
-    cache: FxHashMap<String, T>,
+pub struct ToolManager {
+    cache: FxHashMap<String, CachedTool>,
     runtime: Runtime, // Default workspace version
 }
 
-impl<T: RuntimeTool + Debug> ToolManager<T> {
+impl ToolManager {
     pub fn new(runtime: Runtime) -> Self {
         ToolManager {
             cache: FxHashMap::default(),
@@ -17,18 +20,18 @@ impl<T: RuntimeTool + Debug> ToolManager<T> {
         }
     }
 
-    pub fn get(&self) -> Result<&T, ToolchainError> {
+    pub fn get(&self) -> Result<&CachedTool, ToolchainError> {
         self.get_for_runtime(&self.runtime)
     }
 
-    pub fn get_for_runtime(&self, runtime: &Runtime) -> Result<&T, ToolchainError> {
+    pub fn get_for_runtime(&self, runtime: &Runtime) -> Result<&CachedTool, ToolchainError> {
         match &runtime {
             Runtime::Node(version) => self.get_for_version(&version.0),
             _ => Err(ToolchainError::UnsupportedRuntime(runtime.to_owned())),
         }
     }
 
-    pub fn get_for_version(&self, version: &str) -> Result<&T, ToolchainError> {
+    pub fn get_for_version(&self, version: &str) -> Result<&CachedTool, ToolchainError> {
         if !self.has(version) {
             return Err(ToolchainError::MissingTool(format!(
                 "{} v{}",
@@ -43,7 +46,7 @@ impl<T: RuntimeTool + Debug> ToolManager<T> {
         self.cache.contains_key(version)
     }
 
-    pub fn register(&mut self, tool: T, root: bool) {
+    pub fn register(&mut self, tool: CachedTool, root: bool) {
         // Nothing exists in the cache yet, so this tool must be the top-level
         // workspace tool. If so, update the default version within the platform.
         if self.cache.is_empty() && root {
@@ -67,7 +70,7 @@ impl<T: RuntimeTool + Debug> ToolManager<T> {
         last_versions: &mut FxHashMap<String, String>,
     ) -> Result<u8, ToolchainError> {
         match self.cache.get_mut(version) {
-            Some(cache) => cache.setup(last_versions).await,
+            Some(cache) => Ok(cache.setup(last_versions).await?),
             None => Err(ToolchainError::MissingTool(self.runtime.to_string())),
         }
     }
