--- crates/core/toolchain/src/tools/mod.rs
@@ -1,4 +1,0 @@
-pub mod node;
-pub mod npm;
-pub mod pnpm;
-pub mod yarn;
