+++ crates/proto/core/src/downloader.rs
@@ -1,6 +1,6 @@
-use crate::errors::ProtoError;
 use crate::resolver::Resolvable;
 use log::trace;
+use proto_error::ProtoError;
 use std::fs::File;
 use std::io;
 use std::path::{Path, PathBuf};
