+++ crates/proto/core/src/executor.rs
@@ -1,4 +1,4 @@
-use crate::errors::ProtoError;
+use proto_error::ProtoError;
 use std::path::Path;
 
 #[async_trait::async_trait]
