+++ crates/proto/core/src/lib.rs
@@ -1,7 +1,6 @@
 mod describer;
 mod detector;
 mod downloader;
-mod errors;
 mod executor;
 mod helpers;
 mod installer;
@@ -12,11 +11,11 @@ pub use async_trait::async_trait;
 pub use describer::*;
 pub use detector::*;
 pub use downloader::*;
-pub use errors::*;
 pub use executor::*;
 pub use helpers::*;
 pub use installer::*;
 pub use lenient_semver::Version;
+pub use proto_error::ProtoError;
 pub use resolver::*;
 use std::fs;
 use std::path::{Path, PathBuf};
