+++ crates/proto/error/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "proto_error"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+thiserror = { workspace = true }
